files refactoring and add navigation rail
