+++ lib/pages/parameter_page.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class ParameterPage extends StatelessWidget {
+  const ParameterPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          Text("QSfssd"),
+        ],
+      ),
+    );
+  }
+}
